requirements.txt

fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
pydantic>=2.7.0


huggingface-hub>=0.23.0
transformers>=4.41.0
requests>=2.32.0



google-genai>=0.5.0

ddgs>=9.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

tiktoken>=0.7.0

# ── NLP Metrics ──────────────────────────────
nltk>=3.8.1

# ── NLI Model (DeBERTa-v3 trained on MNLI) ──
# Replaces bert-score with proper NLI-based verification
datasets>=2.19.0
accelerate>=0.30.0
evaluate>=0.4.0
scikit-learn>=1.4.0
torch>=2.2.0
numpy>=1.26.0

# ── Utilities ────────────────────────────────
aiofiles>=23.2.1
tenacity>=8.3.0


protobuf>=5.28.0,<6.0.0


sympy>=1.13.0


tqdm>=4.66.0
pandas>=2.2.0

